Move bicycle admin pages to Bicycles folder
